Updated MyDate.java and StockSimulation.java
Modified function {public void SetMyDate(String date)} and {public String GetMyDateString()}.  Also, commented that the private variable, my_date_ should be stored as yyyymmdd.  Also, the file, Stock (1).java, needs to be deleted as it does not run correctly since the filename needs to be the same as the class name.